Tidy up cli files
