+++ oasislmf/cli/api.py
@@ -1,6 +1,5 @@
 from argparse import RawDescriptionHelpFormatter
-from .base import OasisBaseCommand
-from .computation_command import OasisComputationCommand
+from .command import OasisBaseCommand, OasisComputationCommand
 
 class ListApiCmd(OasisComputationCommand):
     """
