+++ oasislmf/cli/test.py
@@ -1,23 +1,15 @@
-import oasislmf.cli.base
 import tempfile
 import os
 
 from argparsetree import BaseCommand
 
 from ..validation.model_data import csv_validity_test
-
 from ..utils.path import as_path
-
-from ..utils.exceptions import OasisException
+from ..utils.inputs import InputValues
 
 from argparse import RawDescriptionHelpFormatter
-from .computation_command import OasisComputationCommand
-from .base import OasisBaseCommand
-from .inputs import InputValues
+from .command import OasisBaseCommand, OasisComputationCommand
 
-from ..manager import OasisManager as om
-
-# ! Handle multiple tests
 
 
 class ModelValidationCmd(OasisBaseCommand):
