+++ oasislmf/utils/inputs.py
